more orgnized scripts
